server/requirements.txt
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.3.0
certifi==2019.11.28
chardet==3.0.4
click==8.1.8
dbus-python==1.2.16
distro-info==0.23+ubuntu1.1
dnspython==2.6.1
exceptiongroup==1.2.2
fastapi==0.116.1
git-filter-repo==2.47.0
h11==0.14.0
idna==2.8
motor==3.6.1
pydantic==2.10.6
pydantic-core==2.27.2
PyGObject==3.36.0
PyJWT==2.9.0
pymongo==4.9.2
python-apt==2.0.1+ubuntu0.20.4.1
python-dotenv==1.0.1
requests==2.22.0
requests-unixsocket==0.2.0
six==1.14.0
sniffio==1.3.1
starlette==0.44.0
typing-extensions==4.13.2
unattended-upgrades==0.1
urllib3==1.25.8
uvicorn==0.33.0
# QuickPoll

QuickPoll is a full-stack web application that allows users to create, vote, and view polls in real time. Built with a modern stack: **React + Redux Toolkit + TypeScript + Vite** on the frontend, and **FastAPI + MongoDB** on the backend.

## Features

- User registration and authentication (JWT, secure cookies)
- Create, view, and vote on polls
- View latest polls and your own polls
- Responsive, modern UI with Tailwind CSS
- Persistent login with secure session management

## Tech Stack

- **Frontend:** React, Redux Toolkit, TypeScript, Vite, Tailwind CSS
- **Backend:** FastAPI, MongoDB (via Motor), JWT, bcrypt
- **Other:** Axios, Framer Motion, ldrs (loading spinners)

## Getting Started

### Prerequisites

- Node.js (v18+ recommended)
- Python 3.8+
- MongoDB Atlas account (or local MongoDB instance)

### Setup

#### 1. Clone the repository

```sh
git clone https://github.com/yourusername/quickpoll.git
cd quickpoll